updates
Former-commit-id: db8d6210c4af464dce6f012dd964504d05d96a9f
Former-commit-id: d0f8e08ca639bd800782c6324fbbef9bdc6be5cd
Former-commit-id: 5331ec11bd746b06663c4cf514688d5d0d912780

+++ text/text.py
@@ -708,9 +708,10 @@ class Text(object):
 		return passages
 
 	def lines_xml(self):
-		with open(self.fnfn_xml,errors='ignore') as f:
-			for ln in f:
-				yield ln
+		if os.path.exists(self.fnfn_xml):
+			with open(self.fnfn_xml,errors='ignore') as f:
+				for ln in f:
+					yield ln
 
 	def lines_txt(self):
 		if os.path.exists(self.fnfn_txt,errors='ignore'):
